# Faster resorvoir_sampling_single function